Fixed width, rewrote retrieve video as a util rather than as a component.
